Remove require() statements from opencv.js
This allows us to build the UI without needing any polyfills. Ideally
rsbuild should remove these statements as these are only valid when
opencv is run in a node environment, but this does not seem to work.

We will use this fix as a quick workaround for
https://github.com/open-edge-platform/geti/security/dependabot/67

In a follow up PR I'd like to investigate if we can automatically build
opencv without any node references.