Only show specific tiling parameters when mode is custom
Currently the backend always returns the custom tiling parameters even
when they are not being used.
We need this as these values are used to determine the default paramter
values when configuring a new model. So atm we can't remove these
parameters from the API when tiling mode is off.